+++ src/plugins/abbreviationExpansionPlugin.h
@@ -47,7 +47,7 @@ public:
     AbbreviationExpansionPlugin(Profile*, ContextTracker*);
     ~AbbreviationExpansionPlugin();
 
-    virtual Prediction predict(const int) const;
+    virtual Prediction predict(const size_t) const;
 
     virtual void learn();
     virtual void extract();
